# backend/greenride-api-clean/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GreenRide API Documentation\n## Response Codes:\n- 0000: Success - 操作成功\n- 9999: System Error - 系统错误\n- 1000: Parameter Error - 参数错误\n- 1001: Authorization Error - 授权失败\n- 1002: Authentication Expired - 认证过期\n- 2001: User Error - 用户相关错误\n- 4001: Business Error - 业务错误\n- 4002: Data Expired - 数据过期",
        "title": "GreenRide API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/account/delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "软删除用户账户，添加删除标记和时间戳，保留数据但标记为已删除状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "删除账户",
                "parameters": [
                    {
                        "description": "删除请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户修改密码，需要提供旧密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
        "/api/offline": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "司机下线，只有司机类型用户可以调用此接口。下线后司机状态变为离线，车辆状态变为不可用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "司机"
                ],
                "summary": "司机下线",
                "responses": {
                    "200": {
                        "description": "下线成功",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/online": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "司机上线，只有司机类型用户可以调用此接口。上线后司机状态变为在线，车辆状态变为可用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "司机"
                ],
                "summary": "司机上线",
                "responses": {
                    "200": {
                        "description": "上线成功",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "401": {
                        "description": "认证失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
        "/api/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户信息",
                "responses": {
                    "200": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
        "/api/profile/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "按需更新用户个人信息，只更新提供的字段",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新个人信息",
                "parameters": [
                    {
                        "description": "更新请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
        "/api/reset-password": {
            "post": {
                "description": "通过验证码重置密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "重置密码",
                "parameters": [
                    {
                        "description": "重置密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "重置失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
        "/api/vehicle": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前用户的车辆信息，司机返回绑定车辆，车主返回拥有的车辆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "车辆"
                ],
                "summary": "获取用户车辆信息",
                "parameters": [
                    {
                        "description": "查询请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UserVehicleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/protocol.Result"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "description": "新密码",
                    "type": "string"
                },
                "old_password": {
                    "description": "旧密码",
                    "type": "string"
                }
            }
        },
        "handlers.DeleteAccountRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "description": "当前密码，用于确认",
                    "type": "string"
                },
                "reason": {
                    "description": "删除原因（可选）",
                    "type": "string"
                }
            }
        },
...
        "handlers.ResetPasswordRequest": {
            "type": "object",
            "required": [
                "code",
                "method",
                "new_password",
                "user_type"
            ],
            "properties": {
                "code": {
                    "description": "验证码",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "method": {
                    "description": "email, phone",
                    "type": "string",
                    "enum": [
                        "email",
                        "phone"
                    ]
                },
                "new_password": {
                    "description": "新密码",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "user_type": {
                    "description": "用户类型",
                    "type": "string",
                    "enum": [
                        "passenger",
                        "driver"
                    ]
                }
            }
        },
...
        "handlers.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "birthday": {
                    "type": "integer"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "postal_code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "handlers.UserVehicleRequest": {
            "type": "object"
        },
...
        "protocol.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}